ml-self-studio/requirements.txt
opencv-python==4.11.0.86
opencv-contrib-python==4.11.0.86
mediapipe==0.10.21
pillow==11.3.0
pyautogui==0.9.54
pygetwindow==0.0.9
numpy==1.26.4
pywin32==306
ultralytics>=8.0.0

# TensorFlow untuk training EfficientNet-B0
tensorflow>=2.13.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# Web UI
flask>=3.0.0
